refactoring(ContactPage): apply new page building format
